Updated basic sign-in spec
- Externalised login/logout cmds
